Changed Admin Config to use long
Switched the TelegramAdmin config option to use long instead of casting to int

+++ src/main/java/com/gmail/leeyi45/PlayerListPlugin/pluginMain/Config.java
@@ -26,5 +26,5 @@ public class Config
         config.set("telegram.token", arg);
     }
 
-    public static int getTelegramAdmin() { return config.getInt("telegram.admin", -1); }
+    public static long getTelegramAdmin() { return config.getLong("telegram.admin", -1); }
 }
